--- src/app/routes/addInvoice.ts
@@ -1,7 +1,0 @@
-import { adaptRoute } from 'app/adapters/expressRoutesAdapter';
-import { makeAddInvoiceController } from 'app/factories/addInvoice';
-import { Router } from 'express';
-
-export default (router: Router): void => {
-  router.post('/analyze-invoice', adaptRoute(makeAddInvoiceController()));
-};
