+++ src/app/server.ts
@@ -1,4 +1,4 @@
-import { MongoHelper } from '../infra/db/mongodb/helpers/mongoHelper';
+import { MongoHelper } from '@infra/db/mongodb/helpers/mongoHelper';
 import env from './config/env';
 
 MongoHelper.connect(env.mongoUrl)
