+++ src/data/protocols/addInvoiceRepository.ts
@@ -1,5 +1,5 @@
-import { InvoiceModel } from 'domain/models/invoice';
-import { AddInvoiceModel } from 'domain/usecases/addInvoice';
+import { InvoiceModel } from '@domain/models/invoice';
+import { AddInvoiceModel } from '@domain/usecases/addInvoice';
 
 export interface AddInvoiceRepository {
   add(invoiceData: AddInvoiceModel): Promise<InvoiceModel>;
