+++ src/data/protocols/addUserRepository.ts
@@ -1,5 +1,5 @@
-import { UserModel } from 'domain/models/user';
-import { AddUserModel } from 'domain/usecases/addUser';
+import { UserModel } from '@domain/models/user';
+import { AddUserModel } from '@domain/usecases/addUser';
 
 export interface AddUserRepository {
   add(userData: AddUserModel): Promise<UserModel>;
