+++ src/domain/usecases/addUser.ts
@@ -1,4 +1,4 @@
-import { UserModel } from 'domain/models/user';
+import { UserModel } from '@domain/models/user';
 
 export interface AddUserModel {
   name: string;
