+++ src/presentation/helper/httpHelper.ts
@@ -1,4 +1,4 @@
-import { HttpResponse } from 'presentation/protocols';
+import { HttpResponse } from '@presentation/protocols';
 
 export const ok = (data: any): HttpResponse => ({
   statusCode: 200,
